refactor npu patch
