+++ verl_omni/workers/rollout/vllm_rollout/utils.py
@@ -13,61 +13,19 @@
 # limitations under the License.
 import logging
 import os
-from contextlib import AbstractContextManager, contextmanager, nullcontext
 
 import torch
 from verl.workers.rollout.vllm_rollout.utils import VLLM_LORA_INT_ID, VLLM_LORA_NAME, VLLM_LORA_PATH, set_death_signal
-from vllm.utils.mem_utils import GiB_bytes
 from vllm_omni.diffusion.worker.diffusion_worker import CustomPipelineWorkerExtension
 
 from verl_omni.utils.vllm_omni import OmniTensorLoRARequest, VLLMOmniHijack
+from verl_omni.workers.rollout.vllm_rollout.npu_utils import NPUColocateWorkerMixin
 
 logger = logging.getLogger(__file__)
 logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "WARN"))
 
 
-def _is_npu_platform() -> bool:
-    try:
-        from vllm.platforms import current_platform
-
-        return current_platform.device_type == "npu"
-    except Exception:
-        return False
-
-
-def _get_npu_memory_allocator():
-    from vllm_ascend.device_allocator.camem import CaMemAllocator
-
-    return CaMemAllocator.get_instance()
-
-
-@contextmanager
-def _skip_diffusers_npu_empty_cache():
-    try:
-        from diffusers.models import modeling_utils
-        from diffusers.utils import torch_utils
-    except Exception:
-        yield
-        return
-
-    original_modeling_empty_cache = modeling_utils.empty_device_cache
-    original_torch_empty_cache = torch_utils.empty_device_cache
-
-    def empty_device_cache(device_type: str | None = None):
-        if device_type is None or device_type == "npu":
-            return
-        return original_torch_empty_cache(device_type)
-
-    modeling_utils.empty_device_cache = empty_device_cache
-    torch_utils.empty_device_cache = empty_device_cache
-    try:
-        yield
-    finally:
-        modeling_utils.empty_device_cache = original_modeling_empty_cache
-        torch_utils.empty_device_cache = original_torch_empty_cache
-
-
-class vLLMOmniColocateWorkerExtension(CustomPipelineWorkerExtension):
+class vLLMOmniColocateWorkerExtension(NPUColocateWorkerMixin, CustomPipelineWorkerExtension):
     """
     The class for vLLM-Omni's worker to inherit from, in the colocate setting.
     By defining an extension class, the code can work no matter what is
@@ -78,6 +36,7 @@ class vLLMOmniColocateWorkerExtension(CustomPipelineWorkerExtension):
 
     Feature support:
     1. LoRA
+    2. NPU (Ascend) memory-pool, sleep, and wake_up — via NPUColocateWorkerMixin
     """
 
     def __new__(cls, **kwargs):
@@ -87,70 +46,6 @@ class vLLMOmniColocateWorkerExtension(CustomPipelineWorkerExtension):
         VLLMOmniHijack.hijack()
 
         return super().__new__(cls)
-
-    def _maybe_get_memory_pool_context(self, tag: str) -> AbstractContextManager:
-        if not _is_npu_platform():
-            return super()._maybe_get_memory_pool_context(tag)
-
-        if not self.od_config.enable_sleep_mode:
-            return nullcontext()
-
-        allocator = _get_npu_memory_allocator()
-        if tag == "weights":
-            assert allocator.get_current_usage() == 0, "Sleep mode can only be used for one instance per process."
-
-        @contextmanager
-        def npu_memory_pool_context():
-            with _skip_diffusers_npu_empty_cache(), allocator.use_memory_pool(tag=tag):
-                yield
-
-        return npu_memory_pool_context()
-
-    def sleep(self, level: int = 1) -> bool:
-        if not _is_npu_platform():
-            return super().sleep(level)
-
-        free_bytes_before_sleep = None
-        try:
-            free_bytes_before_sleep = torch.npu.mem_get_info()[0]
-        except Exception:
-            pass
-
-        if level == 2 and self.model_runner is not None:
-            model = self.model_runner.pipeline
-            self._sleep_saved_buffers = {name: buffer.cpu().clone() for name, buffer in model.named_buffers()}
-
-        allocator = _get_npu_memory_allocator()
-        allocator.sleep(offload_tags=("weights",) if level == 1 else tuple())
-
-        if free_bytes_before_sleep is not None:
-            try:
-                free_bytes_after_sleep, total = torch.npu.mem_get_info()
-                freed_bytes = free_bytes_after_sleep - free_bytes_before_sleep
-                used_bytes = total - free_bytes_after_sleep
-                logger.info(
-                    "Sleep mode freed %.2f GiB memory, %.2f GiB memory is still in use.",
-                    freed_bytes / GiB_bytes,
-                    used_bytes / GiB_bytes,
-                )
-            except Exception:
-                pass
-        return True
-
-    def wake_up(self, tags: list[str] | None = None) -> bool:
-        if not _is_npu_platform():
-            return super().wake_up(tags)
-
-        allocator = _get_npu_memory_allocator()
-        allocator.wake_up(tags=tags)
-
-        if len(self._sleep_saved_buffers) and self.model_runner is not None:
-            model = self.model_runner.pipeline
-            for name, buffer in model.named_buffers():
-                if name in self._sleep_saved_buffers:
-                    buffer.data.copy_(self._sleep_saved_buffers[name].data)
-            self._sleep_saved_buffers = {}
-        return True
 
     def update_weights_from_ipc(self, peft_config: dict = None, base_sync_done=False, use_shm: bool = False):
         """Update the weights of the rollout model."""
